requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
langchain==0.1.0
langchain-community==0.0.12
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
